@@ -6,12 +6,12 @@ import com.aizuda.easy.security.filter.wrapper.RepWrapper;
 import com.aizuda.easy.security.filter.wrapper.ReqWrapper;
 import com.aizuda.easy.security.properties.SecurityProperties;
 import com.aizuda.easy.security.util.LocalUtil;
+import jakarta.servlet.*;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import javax.servlet.*;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;
 
