@@ -10,16 +10,16 @@ import com.aizuda.easy.security.exp.impl.BasicException;
 import com.aizuda.easy.security.handler.AbstractFunctionHandler;
 import com.aizuda.easy.security.handler.ReqFunctionHandler;
 import com.aizuda.easy.security.util.LocalUtil;
+import jakarta.servlet.http.HttpServletRequest;
 import org.springframework.stereotype.Component;
 
-import javax.servlet.http.HttpServletRequest;
 import java.util.List;
 
 @Component
 public class AuthorizationHandler extends AbstractFunctionHandler implements ReqFunctionHandler {
 
     @Override
-    public String exec(HttpServletRequest request,String json) throws BasicException {
+    public String exec(HttpServletRequest request, String json) throws BasicException {
         LocalEntity localEntity = LocalUtil.getLocalEntity();
         if(localEntity.getSpecial() || localEntity.getProject() || !properties.getAuthorizeEnable()){
             return json;
